requirements.txt
torch>=2.0.0
numpy>=1.24.0
sentence-transformers>=2.2.0
openai>=1.0.0  # 用于调用本地Qwen3 API
networkx>=3.0  # 用于DAG处理
pyyaml>=6.0
tqdm>=4.65.0
matplotlib>=3.7.0
pandas>=2.0.0